+++ ml-agents/mlagents/trainers/tests/test_trainer_util.py
@@ -86,9 +86,8 @@ def dummy_bad_config():
 def test_initialize_trainer_parameters_override_defaults(
     BrainParametersMock, dummy_config_with_override
 ):
-    summaries_dir = "test_dir"
     run_id = "testrun"
-    model_path = "model_dir"
+    output_path = "model_dir"
     keep_checkpoints = 1
     train_model = True
     load_model = False
@@ -97,8 +96,7 @@ def test_initialize_trainer_parameters_override_defaults(
 
     base_config = dummy_config_with_override
     expected_config = base_config["default"]
-    expected_config["summary_path"] = f"{run_id}_testbrain"
-    expected_config["model_path"] = model_path + "/testbrain"
+    expected_config["output_path"] = output_path + "/testbrain"
     expected_config["keep_checkpoints"] = keep_checkpoints
 
     # Override value from specific brain config
@@ -122,9 +120,8 @@ def test_initialize_trainer_parameters_override_defaults(
     with patch.object(PPOTrainer, "__init__", mock_constructor):
         trainer_factory = trainer_util.TrainerFactory(
             trainer_config=base_config,
-            summaries_dir=summaries_dir,
             run_id=run_id,
-            model_path=model_path,
+            output_path=output_path,
             keep_checkpoints=keep_checkpoints,
             train_model=train_model,
             load_model=load_model,
@@ -144,9 +141,8 @@ def test_initialize_ppo_trainer(BrainParametersMock, dummy_config):
     brain_params_mock = BrainParametersMock()
     BrainParametersMock.return_value.brain_name = "testbrain"
     external_brains = {"testbrain": BrainParametersMock()}
-    summaries_dir = "test_dir"
     run_id = "testrun"
-    model_path = "model_dir"
+    output_path = "results_dir"
     keep_checkpoints = 1
     train_model = True
     load_model = False
@@ -155,8 +151,7 @@ def test_initialize_ppo_trainer(BrainParametersMock, dummy_config):
 
     base_config = dummy_config
     expected_config = base_config["default"]
-    expected_config["summary_path"] = f"{run_id}_testbrain"
-    expected_config["model_path"] = model_path + "/testbrain"
+    expected_config["output_path"] = output_path + "/testbrain"
     expected_config["keep_checkpoints"] = keep_checkpoints
 
     def mock_constructor(
@@ -173,9 +168,8 @@ def test_initialize_ppo_trainer(BrainParametersMock, dummy_config):
     with patch.object(PPOTrainer, "__init__", mock_constructor):
         trainer_factory = trainer_util.TrainerFactory(
             trainer_config=base_config,
-            summaries_dir=summaries_dir,
             run_id=run_id,
-            model_path=model_path,
+            output_path=output_path,
             keep_checkpoints=keep_checkpoints,
             train_model=train_model,
             load_model=load_model,
@@ -192,9 +186,8 @@ def test_initialize_ppo_trainer(BrainParametersMock, dummy_config):
 def test_initialize_invalid_trainer_raises_exception(
     BrainParametersMock, dummy_bad_config
 ):
-    summaries_dir = "test_dir"
     run_id = "testrun"
-    model_path = "model_dir"
+    output_path = "results_dir"
     keep_checkpoints = 1
     train_model = True
     load_model = False
@@ -206,9 +199,8 @@ def test_initialize_invalid_trainer_raises_exception(
     with pytest.raises(TrainerConfigError):
         trainer_factory = trainer_util.TrainerFactory(
             trainer_config=bad_config,
-            summaries_dir=summaries_dir,
             run_id=run_id,
-            model_path=model_path,
+            output_path=output_path,
             keep_checkpoints=keep_checkpoints,
             train_model=train_model,
             load_model=load_model,
@@ -223,9 +215,8 @@ def test_initialize_invalid_trainer_raises_exception(
     with pytest.raises(TrainerConfigError):
         trainer_factory = trainer_util.TrainerFactory(
             trainer_config=bad_config,
-            summaries_dir=summaries_dir,
             run_id=run_id,
-            model_path=model_path,
+            output_path=output_path,
             keep_checkpoints=keep_checkpoints,
             train_model=train_model,
             load_model=load_model,
@@ -240,9 +231,8 @@ def test_initialize_invalid_trainer_raises_exception(
     with pytest.raises(UnityTrainerException):
         trainer_factory = trainer_util.TrainerFactory(
             trainer_config=bad_config,
-            summaries_dir=summaries_dir,
             run_id=run_id,
-            model_path=model_path,
+            output_path=output_path,
             keep_checkpoints=keep_checkpoints,
             train_model=train_model,
             load_model=load_model,
@@ -270,9 +260,8 @@ def test_handles_no_default_section(dummy_config):
 
     trainer_factory = trainer_util.TrainerFactory(
         trainer_config=no_default_config,
-        summaries_dir="test_dir",
         run_id="testrun",
-        model_path="model_dir",
+        output_path="output_path",
         keep_checkpoints=1,
         train_model=True,
         load_model=False,
@@ -299,9 +288,8 @@ def test_raise_if_no_config_for_brain(dummy_config):
 
     trainer_factory = trainer_util.TrainerFactory(
         trainer_config=bad_config,
-        summaries_dir="test_dir",
         run_id="testrun",
-        model_path="model_dir",
+        output_path="output_path",
         keep_checkpoints=1,
         train_model=True,
         load_model=False,
@@ -339,33 +327,27 @@ you:
 
 
 def test_existing_directories(tmp_path):
-    model_path = os.path.join(tmp_path, "runid")
-    # Unused summary path
-    summary_path = os.path.join(tmp_path, "runid")
+    output_path = os.path.join(tmp_path, "runid")
     # Test fresh new unused path - should do nothing.
-    trainer_util.handle_existing_directories(model_path, summary_path, False, False)
+    trainer_util.handle_existing_directories(output_path, False, False)
     # Test resume with fresh path - should throw an exception.
     with pytest.raises(UnityTrainerException):
-        trainer_util.handle_existing_directories(model_path, summary_path, True, False)
+        trainer_util.handle_existing_directories(output_path, True, False)
 
     # make a directory
-    os.mkdir(model_path)
+    os.mkdir(output_path)
     # Test try to train w.o. force, should complain
     with pytest.raises(UnityTrainerException):
-        trainer_util.handle_existing_directories(model_path, summary_path, False, False)
+        trainer_util.handle_existing_directories(output_path, False, False)
     # Test try to train w/ resume - should work
-    trainer_util.handle_existing_directories(model_path, summary_path, True, False)
+    trainer_util.handle_existing_directories(output_path, True, False)
     # Test try to train w/ force - should work
-    trainer_util.handle_existing_directories(model_path, summary_path, False, True)
+    trainer_util.handle_existing_directories(output_path, False, True)
 
     # Test initialize option
     init_path = os.path.join(tmp_path, "runid2")
     with pytest.raises(UnityTrainerException):
-        trainer_util.handle_existing_directories(
-            model_path, summary_path, False, True, init_path
-        )
+        trainer_util.handle_existing_directories(output_path, False, True, init_path)
     os.mkdir(init_path)
     # Should pass since the directory exists now.
-    trainer_util.handle_existing_directories(
-        model_path, summary_path, False, True, init_path
-    )
+    trainer_util.handle_existing_directories(output_path, False, True, init_path)